requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.0
pillow==10.3.0
